+++ app/src/components/items/uv/universalViewer.tsx
@@ -1,0 +1,166 @@
+"use client";
+import {
+  useUniversalViewer,
+  useEvent,
+} from "../../../hooks/useUniversalViewer";
+import React, { useEffect, useMemo, useRef } from "react";
+import { IIIFEvents as BaseEvents, IIIFURLAdapter } from "universalviewer";
+
+export type UniversalViewerProps = {
+  config?: any;
+  manifestId: string;
+  canvasIndex?: number;
+  onChangeCanvas?: (manifest: string, canvas: string) => void;
+  onChangeManifest?: (manifest: string) => void;
+};
+
+// https://codesandbox.io/p/sandbox/uv-nextjs-example-239ff5?file=%2Fcomponents%2FUniversalViewer.tsx%3A39%2C1-49%2C8
+const UniversalViewer: React.FC<UniversalViewerProps> = React.memo(
+  ({ manifestId, canvasIndex, onChangeCanvas, config }) => {
+    console.log("canvasIndex in UniversalViewer component is: ", canvasIndex);
+    const ref = useRef<HTMLDivElement>(null);
+    const lastIndex = useRef<number>();
+    const options = useMemo(
+      () => ({
+        manifest: manifestId,
+        canvasIndex: canvasIndex || 0,
+        embedded: true,
+        // footerPanelEnabled: false,
+        // options: {
+        //   "headerPanel": {
+        //     "options": {
+        //       "centerOptionsEnabled": true,
+        //       "localeToggleEnabled": false,
+        //       "settingsButtonEnabled": false
+        //     }
+        //   },
+        // }
+        //UVConfig.options
+        // config: config || UVConfig, //{},
+      }),
+      []
+    );
+
+    const uv = useUniversalViewer(ref, options);
+
+    useEffect(() => {
+      console.log("canvasIndex in useEffect is: ", canvasIndex);
+
+      if (uv) {
+        // override config using an inline json object
+        uv.on("configure", function ({ config, cb }) {
+          cb(
+            {
+              options: {
+                footerPanelEnabled: true,
+                pagingEnabled: true,
+                pagingHeaderPanel: true,
+                pagingOptionEnabled: true,
+                modules: {
+                  headerPanel: {
+                    options: {
+                      centerOptionsEnabled: true,
+                      localeToggleEnabled: false,
+                      settingsButtonEnabled: false,
+                    },
+                  },
+                  pagingHeaderPanel: {
+                    options: {
+                      autoCompleteBoxEnabled: true,
+                      autocompleteAllowWords: false,
+                      galleryButtonEnabled: false,
+                      imageSelectionBoxEnabled: false,
+                      pageModeEnabled: false,
+                      pagingToggleEnabled: true,
+                    },
+                    content: {
+                      close: "Close",
+                      emptyValue: "Please enter a value",
+                      first: "First",
+                      firstImage: "First Image",
+                      firstPage: "First Page",
+                      folio: "Folio",
+                      gallery: "Gallery",
+                      go: "Go",
+                      help: "Help",
+                      image: "Image",
+                      last: "Last",
+                      lastImage: "Last Image",
+                      lastPage: "Last Page",
+                      next: "Next",
+                      nextImage: "Next Image",
+                      nextPage: "Next Page",
+                      of: "of {0}",
+                      oneUp: "Single page view",
+                      page: "Page",
+                      pageSearchLabel: "Search by Page Number",
+                      previous: "Previous",
+                      previousImage: "Previous Image",
+                      previousPage: "Previous Page",
+                      settings: "Settings",
+                      twoUp: "Two page view",
+                    },
+                  },
+                  shareDialogue: {
+                    options: {
+                      embedTemplate:
+                        '<iframe src="{0}" width="{1}" height="{2}" allowfullscreen frameborder="0"></iframe>',
+                      instructionsEnabled: false,
+                      shareFrameEnabled: true,
+                      shareManifestsEnabled: true,
+                    },
+                    content: {
+                      customSize: "custom",
+                      embed: "Embed",
+                      embedInstructions:
+                        "To embed this item in your own website, copy and paste the code below.",
+                      height: "Height",
+                      iiif: "IIIF Manifest",
+                      share: "Share",
+                      shareInstructions:
+                        "To share this item, copy the URL below.",
+                      size: "Size:",
+                      width: "Width",
+                    },
+                  },
+                },
+              },
+            },
+            [uv]
+          );
+        });
+      }
+    });
+
+    useEvent(uv, BaseEvents.CANVAS_INDEX_CHANGE, (i) => {
+      console.log("canvasIndex in useEvent is: ", canvasIndex);
+
+      if (onChangeCanvas) {
+        console.log("canvasIndex in useEvent onChangeCanvas is: ", canvasIndex);
+
+        if (lastIndex.current !== i) {
+          console.log("lastIndex is", lastIndex);
+          const canvas = (uv as any)?.extension?.helper.getCanvasByIndex(i);
+          if (canvas) {
+            lastIndex.current = i;
+            onChangeCanvas(manifestId, canvas.id);
+          }
+        }
+      }
+    });
+
+    return (
+      <>
+        <link
+          rel="stylesheet"
+          href="https://cdn.jsdelivr.net/npm/universalviewer@4.0.0-pre.66/dist/esm/index.css"
+        />
+        <div className="uv" style={{ height: 800 }} ref={ref} />
+      </>
+    );
+  }
+);
+
+UniversalViewer.displayName = "UniversalViewer";
+
+export default UniversalViewer;